+++ src/Services/Airline.Identity/src/Identity/Identity/Models/ApplicationUser.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Identity;
 
 namespace Identity.Identity.Models;
 
-public class ApplicationUser : IdentityUser<int>
+public class ApplicationUser : IdentityUser<long>
 {
     public string FirstName { get; set; }
     public string LastName { get; set; }
